refactor: updated structure
