+++ bungeecord/build.gradle
@@ -1,0 +1,8 @@
+repositories {
+    maven { url 'https://oss.sonatype.org/content/repositories/snapshots/' }
+}
+
+dependencies {
+    implementation project(':core')
+    compileOnly 'net.md-5:bungeecord-api:1.20-R0.1-SNAPSHOT'
+}
