+++ bungeecord/src/main/java/gg/enes/moderation/bungeecord/ModerationTools.java
@@ -1,0 +1,4 @@
+package gg.enes.moderation.bungeecord;
+
+public class ModerationTools {
+}
