+++ core/src/main/java/gg/enes/moderation/core/Core.java
@@ -1,0 +1,4 @@
+package gg.enes.moderation.core;
+
+public class Core {
+}
