+++ core/src/main/java/gg/enes/moderation/core/caching/CacheManager.java
@@ -1,0 +1,9 @@
+package gg.enes.moderation.core.caching;
+
+public interface CacheManager {
+    <K, V> void set(K key, V value);
+
+    <K, V> V get(K key, Class<V> type);
+
+    <K> void del(K key);
+}
