+++ core/src/main/java/gg/enes/moderation/core/entity/annotations/Column.java
@@ -1,0 +1,10 @@
+package gg.enes.moderation.core.entity.annotations;
+
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+
+@Retention(RetentionPolicy.RUNTIME)
+public @interface Column {
+    String name();
+    boolean primary() default false;
+}
