+++ core/src/main/java/gg/enes/moderation/core/entity/annotations/Table.java
@@ -1,0 +1,9 @@
+package gg.enes.moderation.core.entity.annotations;
+
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+
+@Retention(RetentionPolicy.RUNTIME)
+public @interface Table {
+    String name();
+}
