+++ core/src/main/java/gg/enes/moderation/core/repository/BaseRepository.java
@@ -1,0 +1,4 @@
+package gg.enes.moderation.core.repository;
+
+public abstract class BaseRepository {
+}
