+++ core/src/main/java/gg/enes/moderation/core/repository/UserRepository.java
@@ -1,0 +1,4 @@
+package gg.enes.moderation.core.repository;
+
+public class UserRepository extends BaseRepository {
+}
