--- modules/bungeecord/build.gradle
@@ -1,8 +1,0 @@
-repositories {
-    maven { url 'https://oss.sonatype.org/content/repositories/snapshots/' }
-}
-
-dependencies {
-    implementation project(':modules:core')
-    compileOnly 'net.md-5:bungeecord-api:1.20-R0.1-SNAPSHOT'
-}
