--- modules/bungeecord/src/main/java/gg/enes/moderation/ModerationTools.java
@@ -1,4 +1,0 @@
-package gg.enes.moderation;
-
-public class ModerationTools {
-}
