--- modules/core/src/main/java/gg/enes/core/Core.java
@@ -1,4 +1,0 @@
-package gg.enes.core;
-
-public class Core {
-}
