--- modules/core/src/main/java/gg/enes/core/caching/CacheManager.java
@@ -1,9 +1,0 @@
-package gg.enes.core.caching;
-
-public interface CacheManager {
-    <K, V> void set(K key, V value);
-
-    <K, V> V get(K key, Class<V> type);
-
-    <K> void del(K key);
-}
