--- modules/core/src/main/java/gg/enes/core/entity/annotations/Column.java
@@ -1,10 +1,0 @@
-package gg.enes.core.entity.annotations;
-
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-
-@Retention(RetentionPolicy.RUNTIME)
-public @interface Column {
-    String name();
-    boolean primary() default false;
-}
