--- modules/core/src/main/java/gg/enes/core/entity/annotations/Table.java
@@ -1,9 +1,0 @@
-package gg.enes.core.entity.annotations;
-
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-
-@Retention(RetentionPolicy.RUNTIME)
-public @interface Table {
-    String name();
-}
