--- modules/core/src/main/java/gg/enes/core/repository/BaseRepository.java
@@ -1,4 +1,0 @@
-package gg.enes.core.repository;
-
-public abstract class BaseRepository {
-}
