--- modules/core/src/main/java/gg/enes/core/repository/UserRepository.java
@@ -1,4 +1,0 @@
-package gg.enes.core.repository;
-
-public class UserRepository extends BaseRepository {
-}
