--- modules/paper/build.gradle
@@ -1,8 +1,0 @@
-repositories {
-    maven { url 'https://repo.papermc.io/repository/maven-public/' }
-}
-
-dependencies {
-    implementation project(':modules:core')
-    compileOnly 'io.papermc.paper:paper-api:1.20.4-R0.1-SNAPSHOT'
-}
