--- modules/spigot/build.gradle
@@ -1,8 +1,0 @@
-repositories {
-    maven { url 'https://hub.spigotmc.org/nexus/content/repositories/snapshots/' }
-}
-
-dependencies {
-    implementation project(':modules:core')
-    compileOnly 'org.spigotmc:spigot-api:1.20.4-R0.1-SNAPSHOT'
-}
