--- modules/velocity/build.gradle
@@ -1,9 +1,0 @@
-repositories {
-    maven { url = 'https://repo.papermc.io/repository/maven-public/' }
-}
-
-dependencies {
-    implementation project(':modules:core')
-    compileOnly 'com.velocitypowered:velocity-api:3.3.0-SNAPSHOT'
-    annotationProcessor 'com.velocitypowered:velocity-api:3.3.0-SNAPSHOT'
-}
