+++ paper/src/main/java/gg/enes/moderation/paper/ModerationTools.java
@@ -1,0 +1,4 @@
+package gg.enes.moderation.paper;
+
+public class ModerationTools {
+}
