+++ spigot/src/main/java/gg/enes/moderation/spigot/ModerationTools.java
@@ -1,0 +1,4 @@
+package gg.enes.moderation.spigot;
+
+public class ModerationTools {
+}
