+++ velocity/src/main/java/gg/enes/moderation/velocity/ModerationTools.java
@@ -1,0 +1,4 @@
+package gg.enes.moderation.velocity;
+
+public class ModerationTools {
+}
